update generated files for 1.8.6

+++ libtd-ktx/src/main/java/kotlinx/telegram/coroutines/BasicGroupFunctionsKtx.kt
@@ -4,7 +4,7 @@
 //
 package kotlinx.telegram.coroutines
 
-import kotlin.Int
+import kotlin.Long
 import kotlinx.telegram.core.TelegramFlow
 import org.drinkless.td.libcore.telegram.TdApi
 import org.drinkless.td.libcore.telegram.TdApi.BasicGroup
@@ -12,14 +12,14 @@ import org.drinkless.td.libcore.telegram.TdApi.BasicGroupFullInfo
 
 /**
  * Suspend function, which returns information about a basic group by its identifier. This is an
- * offline request if the current user is not a bot.
+ * offline method if the current user is not a bot.
  *
  * @param basicGroupId Basic group identifier.
  *
  * @return [BasicGroup] Represents a basic group of 0-200 users (must be upgraded to a supergroup to
  * accommodate more than 200 users).
  */
-suspend fun TelegramFlow.getBasicGroup(basicGroupId: Int): BasicGroup =
+suspend fun TelegramFlow.getBasicGroup(basicGroupId: Long): BasicGroup =
     this.sendFunctionAsync(TdApi.GetBasicGroup(basicGroupId))
 
 /**
@@ -29,5 +29,5 @@ suspend fun TelegramFlow.getBasicGroup(basicGroupId: Int): BasicGroup =
  *
  * @return [BasicGroupFullInfo] Contains full information about a basic group.
  */
-suspend fun TelegramFlow.getBasicGroupFullInfo(basicGroupId: Int): BasicGroupFullInfo =
+suspend fun TelegramFlow.getBasicGroupFullInfo(basicGroupId: Long): BasicGroupFullInfo =
     this.sendFunctionAsync(TdApi.GetBasicGroupFullInfo(basicGroupId))
